Simplifying implementation
