+++ inc/endpoints/class-components-endpoint.php
@@ -418,8 +418,9 @@ class Components_Endpoint extends Endpoint {
 	 * @see https://github.com/WordPress/gutenberg/issues/1761
 	 *
 	 * @param string $url Rest URL.
+	 * @return string
 	 */
-	public function fix_rest_url( $url ) {
+	public function fix_rest_url( $url ) : string {
 		$path = wp_parse_url( $url, PHP_URL_PATH );
 		return site_url( $path );
 	}
